token验证，权限管理，手机登陆，验证码获取，代码整理

+++ src/main/java/cn/com/wh/ring/app/bean/pojo/UserPojo.java
@@ -13,7 +13,7 @@ public class UserPojo {
     private Long userId;
     private String account;
     private int accountType;
-    private int userInfoId;
+    private Long userInfoId;
     private Date createTime;
     private String bindAccount;
     private int bindAccountType;
@@ -51,11 +51,11 @@ public class UserPojo {
         this.accountType = accountType;
     }
 
-    public int getUserInfoId() {
+    public Long getUserInfoId() {
         return userInfoId;
     }
 
-    public void setUserInfoId(int userInfoId) {
+    public void setUserInfoId(Long userInfoId) {
         this.userInfoId = userInfoId;
     }
